feat: add truncation logic to duplication and smells issues
